Added in basic Combat
health system for both AI melee, and Player are finished. This means that the player now has a health bar. The player can attack the enemy AI, and finally the AI can finally take his revenge on the player and attack the player killing them. This commit also comes with a respawn function as well for the player.