Added scene, created level/room, made simple key mechanic  Nick committed
Created a level layout with current assets in a new scene. Made key functionality with a basic script that opens a gate when collided with. Used an awkward mix of tilemaps and GameObjects that will be much cleaner in the future now that I have a feel for it. Player still spawns in the middle of the room so I put a key there for now