fixed Errors when disconnecting
Before when the players diconnected, AI would not be able to correctly find anyplayers that where still on the server. This was because the Sync List kept that old player game object inside the list. That old object is now removed. Added in a new class called WhenPlayerLeaves, it was added onto each prefab. Also fixed small bug with damaging players from AI.