chickenWizards ?!
added in range AI, the chicken wizard  has figured out how to cast basic spells at the player. The chicken wizard is weaker then the melee chicken with 8 health. that being said the wizard does not want to be up in the front lines and will run away from the player if they get to close to the chicken. I suggest put these guys with melee chickens as they combo well together. Aside from that I also added onto the enemy spawner so you can choose how many of which AI to spawn.